+++ tests/testthat/test-funs_string-lazy.R
@@ -1233,4 +1233,31 @@ test_that("stringr::str_replace_na works", {
   )
 })
 
+test_that("str_equal() works", {
+  test_df <- data.frame(x = c("\u00e1", "\u2126"), y = c("a\u0301", "\u03A9"))
+  test <- as_polars_lf(test_df)
+
+  expect_equal_lazy(
+    mutate(test, eq = str_equal(x, y)) |>
+      pull(eq),
+    mutate(test_df, eq = str_equal(x, y)) |>
+      pull(eq)
+  )
+
+  expect_warning(
+    mutate(test, eq = str_equal(x, y, ignore_case = TRUE)),
+    "doesn't know how to use some arguments"
+  )
+
+  test_df <- data.frame(x = character(), y = character())
+  test <- as_polars_lf(test_df)
+
+  expect_equal_lazy(
+    mutate(test, eq = str_equal(x, y)) |>
+      pull(eq),
+    mutate(test_df, eq = str_equal(x, y)) |>
+      pull(eq)
+  )
+})
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)